+++ generators/typescript/TsObjectGen.ts
@@ -1,4 +1,5 @@
-﻿
+﻿import Arrays = require("../../../ts-mortar/utils/Arrays");
+
 /**
  * @since 2016-3-20
  */
@@ -33,11 +34,11 @@ module TsObjectGen {
                 else {
                     dstLines.push(objStartStr);
                 }
-                Array.prototype.push.apply(dstLines, propLines);
+                Arrays.addAll(dstLines, propLines);
                 dstLines.push(objEndStr);
             }
             lines.push(objStartStr);
-            Array.prototype.push.apply(lines, propLines);
+            Arrays.addAll(lines, propLines);
             lines.push(objEndStr);
         }
         else {