+++ qunit-tests.ts
@@ -37,3 +37,8 @@ testRunner.run({
     code: "./generators/typescript/TsFieldGen",
     tests: "./test/generators/typescript/TsFieldGenTest.js"
 }, callback);
+
+testRunner.run({
+    code: "./generators/DefaultGenTools",
+    tests: "./test/generators/GenToolsTest.js"
+}, callback);